requirements.txt
Django==1.7.6
Pillow==2.7.0
PyJWT==0.4.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-ckeditor==4.4.7
django-suit-ckeditor==0.0.2
django-toolbelt==0.0.1
gunicorn==19.3.0
oauthlib==0.7.2
psycopg2==2.6
python-openid==2.2.5
python-social-auth==0.2.2
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
static3==0.5.1
wsgiref==0.1.2